requirements.txt
Django==4.0.2
django-environ==0.8.1
djangorestframework==3.13.1
drf-yasg==1.20.0
flake8
flake8-bugbear
flake8-commas
flake8-import-order
flake8-variables-names
openpyxl==3.0.9
pep8-naming
